requirements.txt
# requirements.txt
Pillow==9.4.0
# python_version >= "3.8"
